Lord of The Rings README
Welcome to the Lord of the Rings website! This website is a comprehensive resource for all things related to J.R.R. Tolkien's Middle-earth. The site features a beautiful interface and leverages APIs to provide a rich experience for fans of the books and films.
This website is built using HTML, CSS, and JavaScript, with a responsive design that adapts to different screen sizes. It uses APIs to fetch data from various sources, including a custom API specifically designed for the site.
Check it out : akin-rings.netlify.app